% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{decode}
\alias{decode}
\alias{decode_logical}
\alias{decode_numeric}
\alias{decode_integer}
\alias{decode_character}
\alias{decode_character_simple}
\title{Decodes raw values}
\usage{
decode(x, type, n)

decode_logical(x, ...)

decode_numeric(x, ...)

decode_integer(x, ...)

decode_character(x, n)

decode_character_simple(x, ...)
}
\arguments{
\item{x}{values to decode}

\item{type}{type of encoded values}

\item{n}{number of values in `x`, only used when decoding characters}

\item{...}{ignored}
}
\value{
decoded values
}
\description{
Decodes raw values
}
\section{Functions}{
\itemize{
\item \code{decode_logical()}: decodes logicals

\item \code{decode_numeric()}: decodes numerics

\item \code{decode_integer()}: decodes integers

\item \code{decode_character()}: decodes characters

\item \code{decode_character_simple()}: decodes simple character representation

}}
