% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode}
\alias{encode}
\alias{encode_logical}
\alias{encode_numeric}
\alias{encode_integer}
\alias{encode_character}
\alias{encode_character_simple}
\title{Encodes values}
\usage{
encode(x)

encode_logical(x)

encode_numeric(x)

encode_integer(x)

encode_character(x)

encode_character_simple(x)
}
\arguments{
\item{x}{object to encode}
}
\value{
raw representation of `x`
}
\description{
Encodes values
}
\section{Functions}{
\itemize{
\item \code{encode_logical()}: encodes logicals

\item \code{encode_numeric()}: encodes numerics

\item \code{encode_integer()}: encodes integers

\item \code{encode_character()}: encodes characters

\item \code{encode_character_simple()}: encodes characters into a comma-separated representation

}}
