% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{read_index}
\alias{read_index}
\title{Reads the index of a file}
\usage{
read_index(f)
}
\arguments{
\item{f}{file connection}
}
\value{
index
}
\description{
The index consists of
\enumerate{
 \item number of rows of the data.frame
 \item number of columns of the data.frame
 \item types of the data.frame
 \item indices for each column of the data.frame in the file
 \item column names of the data.frame
 \item the offset where the index starts
}
}
