% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_df}
\alias{save_df}
\title{Saves a data.frame}
\usage{
save_df(df, file_name)
}
\arguments{
\item{df}{data.frame to save}

\item{file_name}{name of file}
}
\value{
nothing
}
\description{
Saves a data.frame
}
