% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{write_header}
\alias{write_header}
\alias{increase_counter}
\alias{write}
\alias{write_column}
\title{Writes data to the data}
\usage{
write_header(file_counter, version = 1)

increase_counter(file_counter, add)

write(x, file_counter)

write_column(x, file_counter)
}
\arguments{
\item{file_counter}{the file counter, contains a file connection and a
counter of how many raws have been written to it}

\item{version}{version, currently not used}

\item{add}{the amount to increase the counter by}

\item{x}{object to write, should be a raw}
}
\value{
`file_counter` with increased counter
}
\description{
Writes data to the data
}
\section{Functions}{
\itemize{
\item \code{write_header()}: writes the header

\item \code{increase_counter()}: increases the counter in the file counter

\item \code{write_column()}: writes a column of a data.frame

}}
