% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{write_index}
\alias{write_index}
\alias{get_types}
\title{Writes the index}
\usage{
write_index(df, file_counter, indices)

get_types(df)
}
\arguments{
\item{df}{the data.frame being saved}

\item{file_counter}{file counter}

\item{indices}{indices of each column in the file}
}
\value{
`file_counter` with increased counter
}
\description{
Writes the index
}
\section{Functions}{
\itemize{
\item \code{get_types()}: numeric encoding of a data.frames types

}}
